Implement engine

+++ tests/variant/main.cpp
@@ -61,9 +61,7 @@ static void testUserType() {
         assert(p.x == 2 && p.y == 3);
     }
 
-    {
-        Variant var1 = Variant::fromValue(uintptr_t(&ss));
-    }
+    { Variant var1 = Variant::fromValue(uint64_t(&ss)); }
 
     assert(Point::g_count() == 0);
 }
